+++ examples.json
@@ -3,7 +3,6 @@
         "test-renderer",
         "camera-navigation",
         "cornell-box",
-        "cornell-box-webgl1",
         "sky-triangle"
     ]
 }
