+++ source/cornell-box/web1_scene.ts
