--- website/cornell-box-webgl1.pug
@@ -1,17 +1,0 @@
-extends partial/example.pug
-
-block vars 
-  - var title   = 'Cornell Box for WebGL 1' 
-  - var author  = ''
-  - var caption = 'ToDo.'
-  - var source  = 'cornell-box-webgl1.js'
-
-
-include partial/code.pug
-
-block breakdown 
-  section.container
-    h4 Example Breadown
-    p This should describe the example in much more detail, e.g., refering to relevant code snippets and resources
-  
-    +code('example-code', 'const example.code.for.breakdown = here;', 'Example code for source break down:')
